Fix 2FA auth
Login was failing when 2FA was enabled because the `ProtonAPI2FANeeded`
exception was not handled in the right place.

Previously, the `proton.vpn.session.VPNSession.authenticate` method was both
authenticating and refreshing the VPN info. When 2FA is required,
the call to refresh the VPN info raises a `ProtonAPI2FANeeded` exception,
which was not handled.

Now, `proton.vpn.session.VPNSession.authenticate` has been changed so
that it does only one thing, it authenticates the user. This allows
us to refactor the orchestrator to be more explicit about when the 2FA
exception is expected and what still needs to be done after it happens
(i.e. refresh the vpn info).