[VPNLINUX-848] Fix error saving server list
Some users get an error while saving the server list to disk.
I assume it happens because users remove the cache directory
while the app is running, and then when the next server list
is retrieved from the API, the app cannot store it because
the parent directory where it wants to save it does not exist.

This parent directory was only created the first time it was
accessed. Now, if it doesn't exist, it's created before
trying to save the cache file.