[VPNLINUX-1126] Notify subscribers when state context changes
Sometimes the state machine doesn't change of state but the
state context changes. In these cases, we now notify connection
state subscribers.

For example, this behaviour is needed  when the current port
forwarding lease changes. In this case the state machine
transitions from Connected state (with port lease X) back to
Connected state (with port lease Y) and the app needs to
reflect this state context change by showing the new port
on the UI.