[VPNLINUX-1173] Always load feature flag cache/defaults
This change stems from a bug where when failing to load the server
list cache, feature flags were left unitialized. But then, feature
flags were required to fetch the server list, so the app was left
in a broken state.

Now feature flags are always initialized from cache, or with defaults
if the cache doesn't exist or is invalid.