chore: [VPNLINUX-1310] change versioning strategy to support the CLI
Up until now, we were sending the version of the app package to the REST
API via the x-pm-appversion header. With this changes, we'll send the
version of our api package instead.

The reason for this is that we'll have 2 frontends (GUI and CLI) using
the same api package to do REST API requests.

That's why we bump the vesion from 0 to 4, so that our backend infra
doesn't think the request comes from a very old version of the app.